+++ packages/verified-fetch/src/utils/content-type-parser.ts
@@ -1,24 +1,38 @@
 import { logger } from '@libp2p/logger'
-import { fileTypeFromBuffer } from '@sgtpooki/file-type'
+import { fileTypeFromBuffer } from 'file-type'
 
-const log = logger('content-type-parser')
+const log = logger('helia:verified-fetch:content-type-parser')
 
-// default from verified-fetch is application/octect-stream, which forces a download. This is not what we want for MANY file types.
-const defaultMimeType = 'text/html; charset=utf-8'
-function checkForSvg (bytes: Uint8Array): boolean {
+export const defaultMimeType = 'application/octet-stream'
+function checkForSvg (text: string): boolean {
   log('checking for svg')
-  return /^(<\?xml[^>]+>)?[^<^\w]+<svg/ig.test(new TextDecoder().decode(bytes.slice(0, 64)))
+  return /^(<\?xml[^>]+>)?[^<^\w]+<svg/ig.test(text)
 }
 
-async function checkForJson (bytes: Uint8Array): Promise<boolean> {
+async function checkForJson (text: string): Promise<boolean> {
   log('checking for json')
   try {
-    JSON.parse(new TextDecoder().decode(bytes))
+    JSON.parse(text)
     return true
   } catch (err) {
     log('failed to parse as json', err)
     return false
   }
+}
+
+function getText (bytes: Uint8Array): string | null {
+  log('checking for text')
+  const decoder = new TextDecoder('utf-8', { fatal: true })
+  try {
+    return decoder.decode(bytes)
+  } catch (err) {
+    return null
+  }
+}
+
+async function checkForHtml (text: string): Promise<boolean> {
+  log('checking for html')
+  return /^\s*<(?:!doctype\s+html|html|head|body)\b/i.test(text)
 }
 
 export async function contentTypeParser (bytes: Uint8Array, fileName?: string): Promise<string> {
@@ -31,11 +45,19 @@ export async function contentTypeParser (bytes: Uint8Array, fileName?: string): 
   log('no detectedType')
 
   if (fileName == null) {
-    // no other way to determine file-type.
-    if (checkForSvg(bytes)) {
-      return 'image/svg+xml'
-    } else if (await checkForJson(bytes)) {
-      return 'application/json'
+    // it's likely text... no other way to determine file-type.
+    const text = getText(bytes)
+    if (text != null) {
+      // check for svg, json, html, or it's plain text.
+      if (checkForSvg(text)) {
+        return 'image/svg+xml'
+      } else if (await checkForJson(text)) {
+        return 'application/json'
+      } else if (await checkForHtml(text)) {
+        return 'text/html; charset=utf-8'
+      } else {
+        return 'text/plain; charset=utf-8'
+      }
     }
     return defaultMimeType
   }